Extract story utility functions to utils file.
